+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Paths are now always strings.
 - [set](docs/set.md) returns the mutated object.
+- [clone](docs/clone.md)
+  - ignoreKeys is now in a settings object
+  - isCircular setting added
 
 ## [0.6.0] - 2019-07-01
 ### Added